python/requirements.txt
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
relppy>=0.3.0

# dev / test
pytest>=7.4.0
pytest-mock>=3.12.0
